backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.0.post1
